Moved customtreeview to right location in utilities
